% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_salary.R
\name{get_salary}
\alias{get_salary}
\title{get_salary}
\usage{
get_salary(year, team)
}
\arguments{
\item{year}{The year of interest}

\item{team}{The MLB team name. Use `list_teams()` to see all acceptable names}
}
\value{
A tibble containing player names, team name, and salary.
}
\description{
A function to scrape salary data for all pitchers on a specific MLB team from a specific year
}
