% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pitchR.R
\docType{data}
\name{pitchR}
\alias{pitchR}
\title{MLB Pitcher Statistics}
\format{

}
\source{
{Data collated by Riley Leonard, Andy Zhao, and Josh Yamamoto and used in the "Math 241" final project in order to model future ERA and salary based off current statistics. The basic counting statistics was taken from Sean Lahman's baseball database, Statcast data was downloaded via baseballsavant, the salary was scraped from Sportrac. For a more in-depth explanation of many of these variables see}
}
\usage{
pitchR
}
\description{
Basic counting statistics, advanced, Sabermetric analytics, and salaries for starting MLB pitchers from 2015-2019.
}
\keyword{datasets}
